Update server with routes and controllers  - updated Locs model to match new data and table
