+++ server/routes.js
@@ -2,3 +2,4 @@ var express = require('express');
 var router = express.Router();
 var Promise = require('bluebird');
 
+// Move routes to this file when server.js has too many routes
