+++ sqldb/config.js
@@ -43,6 +43,10 @@ db.knex.schema.hasTable('Locs').then(function (exists) {
       table.uuid('id').primary();
       table.string('name', 255);
       table.timestamps();
+      table.integer('jobcount', 11);
+      table.float('latitude');
+      table.float('longitude');
+      table.float('avg_salary');
     }).then(function (table) {
       console.log('Create Table', table);
     });
@@ -89,8 +93,8 @@ db.knex.schema.hasTable('loc_job').then(function (exists) {
   if (!exists) {
     db.knex.schema.createTable('loc_job', function (table) {
       table.timestamps();
-      table.uuid('Job_rowId').primary();
-      table.uuid('Loc_rowId').primary();
+      table.uuid('Job_rowId').references('id').inTable('Jobs');
+      table.uuid('Loc_rowId').references('id').inTable('Locs');
     }).then(function (table) {
       console.log('Create Table', table);
     });
@@ -101,8 +105,8 @@ db.knex.schema.hasTable('role_job').then(function (exists) {
   if (!exists) {
     db.knex.schema.createTable('role_job', function (table) {
       table.timestamps();
-      table.uuid('Job_rowId').primary();
-      table.uuid('Role_rowId').primary();
+      table.uuid('Job_rowId').references('id').inTable('Jobs');
+      table.uuid('Role_rowId').references('id').inTable('Roles');
     }).then(function (table) {
       console.log('Create Table', table);
     });
@@ -113,8 +117,8 @@ db.knex.schema.hasTable('skill_job').then(function (exists) {
   if (!exists) {
     db.knex.schema.createTable('skill_job', function (table) {
       table.timestamps();
-      table.uuid('Job_rowId').primary();
-      table.uuid('Skill_rowId').primary();
+      table.uuid('Job_rowId').references('id').inTable('Jobs');
+      table.uuid('Skill_rowId').references('id').inTable('Skills');
     }).then(function (table) {
       console.log('Create Table', table);
     });
@@ -125,8 +129,8 @@ db.knex.schema.hasTable('startup_job').then(function (exists) {
   if (!exists) {
     db.knex.schema.createTable('startup_job', function (table) {
       table.timestamps();
-      table.uuid('Job_rowId').primary();
-      table.uuid('Startup_rowId').primary();
+      table.uuid('Job_rowId').references('id').inTable('Jobs');
+      table.uuid('Startup_rowId').references('id').inTable('Startups');
     }).then(function (table) {
       console.log('Create Table', table);
     });
